+++ packages/website/src/components/PlaygroundPage/index.tsx
@@ -2,20 +2,10 @@ import { fromByteArray, toByteArray } from "base64-js";
 import { deflate, inflate } from "pako";
 import { FC, useState } from "react";
 
-import {
-  Button,
-  Dropdown,
-  DropdownMenu,
-  DropdownMenuActionItem,
-} from "@quri/ui";
-
 import { ShareButton } from "./ShareButton";
-import {
-  Version,
-  VersionedPlayground,
-  defaultVersion,
-  versions,
-} from "./VersionedPlayground";
+import { VersionedPlayground } from "./VersionedPlayground";
+import { VersionPicker } from "./VersionPicker";
+import { defaultVersion, Version, versions } from "./versions";
 
 const HASH_PREFIX = "#code=";
 function getHashData() {
@@ -55,34 +45,6 @@ function updateUrl(data: Partial<Data>, version: Version) {
       encodeURIComponent(fromByteArray(compressed))
   );
 }
-
-const VersionPicker: FC<{
-  version: Version;
-  onChange: (newVersion: Version) => void;
-}> = ({ version, onChange }) => {
-  return (
-    <div className="flex gap-2 items-center">
-      <Dropdown
-        render={({ close }) => (
-          <DropdownMenu>
-            {versions.map((version) => (
-              <DropdownMenuActionItem
-                key={version}
-                title={version}
-                onClick={() => {
-                  onChange(version);
-                  close();
-                }}
-              />
-            ))}
-          </DropdownMenu>
-        )}
-      >
-        <Button size="small">{version}</Button>
-      </Dropdown>
-    </div>
-  );
-};
 
 export const PlaygroundPage: FC<{ version: string | null }> = (props) => {
   const hashData = getHashData();
